hubspot_etl/requirements.txt
# HubSpot to Snowflake ETL Pipeline Requirements
# Updated with Python 3.13 compatible versions

# Core dependencies
requests>=2.28.0,<3.0.0
pandas>=2.0.0,<3.0.0
snowflake-connector-python==3.7.1
python-dotenv>=1.0.0,<2.0.0

# Authentication and security - pinned to compatible versions
cryptography==41.0.7

# Data processing
numpy>=1.23.0,<2.0.0

# AI/ML Dependencies
scikit-learn>=1.3.0,<2.0.0
joblib>=1.3.0,<2.0.0
matplotlib>=3.7.0,<4.0.0
seaborn>=0.12.0,<1.0.0

# Additional dependencies for Python 3.13 compatibility
pyOpenSSL==23.3.0
cffi>=1.17.0

# Optional: For enhanced functionality
# Airflow integration (uncomment if using Airflow)
# apache-airflow>=2.5.0,<3.0.0

# Optional: For Prefect integration (uncomment if using Prefect)
# prefect>=2.0.0,<3.0.0

# Development and testing (uncomment for development)
# pytest>=7.0.0,<8.0.0
# pytest-asyncio>=0.18.0,<1.0.0
# black>=22.0.0,<24.0.0
# flake8>=4.0.0,<7.0.0
# mypy>=1.0.0,<2.0.0
